refactor: updated waitUntil types
This reverts commit 269f1d2a0e2b5eb59c8cd31b33831f0284b8fe3f.
